APP09831-transaction-aggregation-service/.github/load-test/scripts/requirements.txt
# Python dependencies for load test scripts
psycopg2-binary==2.9.9

